gest_stock/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
hiredis==3.0.0
idna==3.10
packaging==24.2
paypalrestsdk==1.13.3
psycopg2-binary
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-decouple==3.8
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==11.2.0
typing_extensions==4.12.2
tzdata==2024.2
django-cors-headers==4.6.0
psycopg2-binary
